[IMP] stock_picking_auto_create_lot: lot on print
There could be several ways to print a detailed operation report label.
For example #1922 or even custom report triggered from a report action or
a server action.

To support these different scenarios, a context can be injected to the
report_action method in order to allow forcing the creation of the
detailed operations lots prior to the label printing.

MT-9803